requirements.txt
django-allauth
django-summernote